# docs/openapi.yaml
openapi: 3.0.0
paths:
...
  /admin/products/{productId}/references:
    get:
      description: OWNER, ADMIN, and STAFF can read references for one product.
      operationId: AdminProductReferencesController_findAllForProduct
      parameters:
        - name: productId
          required: true
          in: path
          schema:
            type: string
        - name: page
          required: false
          in: query
          schema:
            default: 1
            example: 1
            type: number
        - name: pageSize
          required: false
          in: query
          schema:
            maximum: 100
            default: 20
            example: 20
            type: number
        - name: search
          required: false
          in: query
          schema:
            example: RF2
            type: string
        - name: isActive
          required: false
          in: query
          schema:
            example: true
            type: boolean
        - name: inStock
          required: false
          in: query
          schema:
            example: true
            type: boolean
      responses:
        "200":
          description: Paginated product references.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductReferenceResponse"
        "401": &a25
          description: Authentication required.
        "403": &a26
          description: Insufficient admin role.
        "404":
          description: Product not found.
      security: &a27
        - bearer: []
      summary: List product references for a product
      tags: &a28
        - Admin Product References
    post:
      description: OWNER and ADMIN can create product references and compatibility
        attributes.
      operationId: AdminProductReferencesController_create
      parameters:
        - name: productId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductReferenceDto"
      responses:
        "200":
          description: Product reference created.
        "401": *a25
        "403": *a26
        "409":
          description: Duplicate reference code, SKU, or barcode.
      security: *a27
      summary: Create product reference
      tags: *a28
  /admin/product-references/{id}:
    get:
      description: OWNER, ADMIN, and STAFF can read product reference details.
      operationId: AdminProductReferencesController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: Product reference details.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductReferenceResponse"
        "401": *a25
        "403": *a26
        "404":
          description: Product reference not found.
      security: *a27
      summary: Get product reference details
      tags: *a28
    patch:
      description: OWNER and ADMIN can update references and replace compatibility
        attributes.
      operationId: AdminProductReferencesController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductReferenceDto"
      responses:
        "200":
          description: Product reference updated.
        "401": *a25
        "403": *a26
        "404":
          description: Product reference not found.
        "409":
          description: Duplicate reference code, SKU, or barcode.
      security: *a27
      summary: Update product reference
      tags: *a28
    delete:
      description: OWNER and ADMIN can soft-deactivate product references.
      operationId: AdminProductReferencesController_deactivate
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: Product reference deactivated.
        "401": *a25
        "403": *a26
        "404":
          description: Product reference not found.
      security: *a27
      summary: Deactivate product reference
      tags: *a28
  /admin/product-references/{id}/stock:
    patch:
      description: OWNER and ADMIN can update stock, reserved quantity, and low-stock
        threshold. This does not reserve or deduct stock automatically.
      operationId: AdminProductReferencesController_updateStock
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateReferenceStockDto"
      responses:
        "200":
          description: Product reference stock updated.
        "401": *a25
        "403": *a26
        "404":
          description: Product reference not found.
      security: *a27
      summary: Update product reference stock counts
      tags: *a28
...
  /packs/{id}:
    get:
      description: Returns one active pack with attributes, items, products, fixed
        references, and useful product references.
      operationId: PacksController_findOne
      parameters:
        - name: id
          required: true
          in: path
          description: Pack ID.
          schema:
            example: 00000000-0000-4000-8000-000000000001
            type: string
      responses:
        "200":
          description: Pack details.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PackResponse"
        "404":
          description: Pack not found.
      summary: Get an active pack by ID
      tags: *a29
  /admin/packs:
    get:
      description: OWNER, ADMIN, and STAFF can read paginated packs with admin metadata.
      operationId: AdminPacksController_findAll
      parameters:
        - name: page
          required: false
          in: query
          schema:
            default: 1
            example: 1
            type: number
        - name: pageSize
          required: false
          in: query
          schema:
            maximum: 100
            default: 20
            example: 20
            type: number
        - name: search
          required: false
          in: query
          schema:
            example: natural
            type: string
        - name: status
          required: false
          in: query
          schema:
            type: string
            enum:
              - DRAFT
              - ACTIVE
              - ARCHIVED
        - name: isActive
          required: false
          in: query
          schema:
            example: true
            type: boolean
        - name: priceMode
          required: false
          in: query
          schema:
            type: string
            enum:
              - FIXED
              - SUM_ITEMS
              - SUM_ITEMS_WITH_DISCOUNT
        - name: minPrice
          required: false
          in: query
          schema:
            example: 200
            type: number
        - name: maxPrice
          required: false
          in: query
          schema:
            example: 350
            type: number
        - name: sortBy
          required: false
          in: query
          schema:
            type: string
            enum:
              - createdAt
              - name
              - fixedPrice
              - priority
        - name: sortOrder
          required: false
          in: query
          schema:
            type: string
            enum:
              - asc
              - desc
      responses:
        "200":
          description: Paginated packs.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PackResponse"
        "401": &a30
          description: Authentication required.
        "403": &a31
          description: Insufficient admin role.
      security: &a32
        - bearer: []
      summary: List admin packs
      tags: &a33
        - Admin Packs
    post:
      description: OWNER and ADMIN can create draft or active packs. Active packs must
        pass item, pricing, and attribute validation.
      operationId: AdminPacksController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePackDto"
      responses:
        "200":
          description: Pack created.
        "401": *a30
        "403": *a31
        "409":
          description: Pack slug already exists.
      security: *a32
      summary: Create pack
      tags: *a33
  /admin/packs/{id}:
    get:
      description: OWNER, ADMIN, and STAFF can read full pack details.
      operationId: AdminPacksController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: Pack details.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PackResponse"
        "401": *a30
        "403": *a31
        "404":
          description: Pack not found.
      security: *a32
      summary: Get admin pack details
      tags: *a33
    patch:
      description: OWNER and ADMIN can update packs. Nested items and attributes are
        replaced transactionally when provided.
      operationId: AdminPacksController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePackDto"
      responses:
        "200":
          description: Pack updated.
        "401": *a30
        "403": *a31
        "404":
          description: Pack not found.
        "409":
          description: Pack slug already exists.
      security: *a32
      summary: Update pack
      tags: *a33
    delete:
      description: OWNER and ADMIN can archive a pack without deleting historical relations.
      operationId: AdminPacksController_archive
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: Pack archived.
        "401": *a30
        "403": *a31
        "404":
          description: Pack not found.
      security: *a32
      summary: Archive pack
      tags: *a33
...
info:
  title: Beauty Pack Recommendation API
  description: Backend API for beauty-pack personalization, recommendations, COD
    orders, and administration
  version: 1.0.0
  contact: {}
...
servers:
  - url: http://localhost:3000
    description: Local development
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
      description: Paste the JWT access token returned by POST /auth/login.
  schemas:
...
    AttributeOptionResponse:
      type: object
      properties:
        id:
          type: string
          example: 00000000-0000-4000-8000-000000000001
        code:
          type: string
          example: MEDIUM
        label:
          type: string
          example: Medium
      required:
        - id
        - code
        - label
...
    QuizAttributeGroupResponse:
      type: object
      properties:
        code:
          type: string
          example: SKIN_COLOR
        name:
          type: string
          example: Skin Color
      required:
        - code
        - name
...
    CategorySummaryResponse:
      type: object
      properties:
        id:
          type: string
          example: 00000000-0000-4000-8000-000000000001
        code:
          type: string
          example: FOUNDATION
        name:
          type: string
          example: Foundation
      required:
        - id
        - code
        - name
    BrandSummaryResponse:
      type: object
      properties:
        id:
          type: string
          example: 00000000-0000-4000-8000-000000000001
        name:
          type: string
          example: Demo Beauty
      required:
        - id
        - name
    AttributeMatchResponse:
      type: object
      properties:
        id:
          type: string
          example: 00000000-0000-4000-8000-000000000001
        matchType:
          type: string
          enum:
            - COMPATIBLE
            - NOT_COMPATIBLE
            - BOOST
          example: COMPATIBLE
        scoreValue:
          type: number
          example: 40
        isHardFilter:
          type: boolean
          example: false
        attributeGroup:
          $ref: "#/components/schemas/QuizAttributeGroupResponse"
        attributeOption:
          $ref: "#/components/schemas/AttributeOptionResponse"
      required:
        - id
        - matchType
        - scoreValue
        - isHardFilter
        - attributeGroup
        - attributeOption
    ProductReferenceResponse:
      type: object
      properties:
        id:
          type: string
          example: 00000000-0000-4000-8000-000000000001
        referenceCode:
          type: string
          example: RF2
        referenceName:
          type: string
          example: Medium Warm
        sku:
          type: object
          example: FOUNDATION-X-RF2
          nullable: true
        barcode:
          type: object
          example: null
          nullable: true
        priceDelta:
          type: number
          example: 0
        stockQuantity:
          type: number
          example: 20
        reservedQuantity:
          type: number
          example: 0
        isActive:
          type: boolean
          example: true
        attributes:
          type: array
          items:
            $ref: "#/components/schemas/AttributeMatchResponse"
      required:
        - id
        - referenceCode
        - referenceName
        - priceDelta
        - stockQuantity
        - reservedQuantity
        - isActive
        - attributes
    ProductResponse:
      type: object
      properties:
        id:
          type: string
          example: 00000000-0000-4000-8000-000000000001
        name:
          type: string
          example: Foundation X
        slug:
          type: string
          example: foundation-x
        basePrice:
          type: number
          example: 120
        currency:
          type: string
          example: MAD
        status:
          type: string
          enum:
            - DRAFT
            - ACTIVE
            - ARCHIVED
          example: ACTIVE
        category:
          $ref: "#/components/schemas/CategorySummaryResponse"
        brand:
          nullable: true
          type: object
          allOf:
            - $ref: "#/components/schemas/BrandSummaryResponse"
        references:
          type: array
          items:
            $ref: "#/components/schemas/ProductReferenceResponse"
      required:
        - id
        - name
        - slug
        - basePrice
        - currency
        - status
        - category
        - references
...
    ReferenceAttributeInputDto:
      type: object
      properties:
        attributeGroupCode:
          type: string
          example: SKIN_COLOR
        attributeOptionCode:
          type: string
          example: MEDIUM
        matchType:
          type: string
          enum:
            - COMPATIBLE
            - NOT_COMPATIBLE
            - BOOST
          example: COMPATIBLE
        scoreValue:
          type: number
          example: 40
        isHardFilter:
          type: boolean
          example: false
          default: false
      required:
        - attributeGroupCode
        - attributeOptionCode
        - matchType
        - scoreValue
    CreateProductReferenceDto:
      type: object
      properties:
        referenceCode:
          type: string
          example: RF2
        referenceName:
          type: string
          example: Medium Warm
        barcode:
          type: object
          example: null
        sku:
          type: object
          example: FOUNDATION-X-RF2
        priceOverride:
          type: object
          example: null
        priceDelta:
          type: number
          example: 0
          default: 0
        imageUrl:
          type: object
          example: https://example.com/reference.jpg
        stockQuantity:
          type: number
          example: 20
          default: 0
        reservedQuantity:
          type: number
          example: 0
          default: 0
        lowStockThreshold:
          type: number
          example: 5
          default: 5
        isDefault:
          type: boolean
          example: false
          default: false
        isActive:
          type: boolean
          example: true
          default: true
        attributes:
          type: array
          items:
            $ref: "#/components/schemas/ReferenceAttributeInputDto"
      required:
        - referenceCode
        - referenceName
    UpdateProductReferenceDto:
      type: object
      properties:
        referenceCode:
          type: string
          example: RF2
        referenceName:
          type: string
          example: Medium Warm
        barcode:
          type: object
          example: null
        sku:
          type: object
          example: FOUNDATION-X-RF2
        priceOverride:
          type: object
          example: null
        priceDelta:
          type: number
          example: 0
          default: 0
        imageUrl:
          type: object
          example: https://example.com/reference.jpg
        stockQuantity:
          type: number
          example: 20
          default: 0
        reservedQuantity:
          type: number
          example: 0
          default: 0
        lowStockThreshold:
          type: number
          example: 5
          default: 5
        isDefault:
          type: boolean
          example: false
          default: false
        isActive:
          type: boolean
          example: true
          default: true
        attributes:
          type: array
          items:
            $ref: "#/components/schemas/ReferenceAttributeInputDto"
    UpdateReferenceStockDto:
      type: object
      properties:
        stockQuantity:
          type: number
          example: 30
        reservedQuantity:
          type: number
          example: 2
        lowStockThreshold:
          type: number
          example: 5
      required:
        - stockQuantity
        - reservedQuantity
        - lowStockThreshold
    PackItemResponse:
      type: object
      properties:
        id:
          type: string
          example: 00000000-0000-4000-8000-000000000001
        selectionMode:
          type: string
          enum:
            - FIXED_REFERENCE
            - AUTO_BEST_REFERENCE
            - CUSTOMER_CHOICE
          example: AUTO_BEST_REFERENCE
        quantity:
          type: number
          example: 1
        isRequired:
          type: boolean
          example: true
        product:
          $ref: "#/components/schemas/ProductResponse"
        productReference:
          nullable: true
          type: object
          allOf:
            - $ref: "#/components/schemas/ProductReferenceResponse"
      required:
        - id
        - selectionMode
        - quantity
        - isRequired
        - product
    PackResponse:
      type: object
      properties:
        id:
          type: string
          example: 00000000-0000-4000-8000-000000000001
        name:
          type: string
          example: Natural Glow Pack
        slug:
          type: string
          example: natural-glow-pack
        priceMode:
          type: string
          enum:
            - FIXED
            - SUM_ITEMS
            - SUM_ITEMS_WITH_DISCOUNT
          example: FIXED
        fixedPrice:
          type: object
          example: 299
          nullable: true
        status:
          type: string
          enum:
            - DRAFT
            - ACTIVE
            - ARCHIVED
          example: ACTIVE
        isActive:
          type: boolean
          example: true
        attributes:
          type: array
          items:
            $ref: "#/components/schemas/AttributeMatchResponse"
        items:
          type: array
          items:
            $ref: "#/components/schemas/PackItemResponse"
      required:
        - id
        - name
        - slug
        - priceMode
        - status
        - isActive
        - attributes
        - items
    PackItemInputDto:
      type: object
      properties:
        productId:
          type: string
          example: 9f72caaa-f55b-4423-b805-911e3e7f61bb
        productReferenceId:
          type: object
          example: 9f72caaa-f55b-4423-b805-911e3e7f61bb
        selectionMode:
          type: string
          enum:
            - FIXED_REFERENCE
            - AUTO_BEST_REFERENCE
            - CUSTOMER_CHOICE
          example: AUTO_BEST_REFERENCE
        quantity:
          type: number
          example: 1
          default: 1
        isRequired:
          type: boolean
          example: true
          default: true
        sortOrder:
          type: number
          example: 1
          default: 0
      required:
        - productId
        - selectionMode
    PackAttributeInputDto:
      type: object
      properties:
        attributeGroupCode:
          type: string
          example: STYLE
        attributeOptionCode:
          type: string
          example: NATURAL
        matchType:
          type: string
          enum:
            - COMPATIBLE
            - NOT_COMPATIBLE
            - BOOST
          example: COMPATIBLE
        scoreValue:
          type: number
          example: 20
        isHardFilter:
          type: boolean
          example: false
          default: false
      required:
        - attributeGroupCode
        - attributeOptionCode
        - matchType
        - scoreValue
    CreatePackDto:
      type: object
      properties:
        name:
          type: string
          example: Natural Glow Pack
        slug:
          type: string
          example: natural-glow-pack
        description:
          type: object
          example: Natural makeup pack
        mainImageUrl:
          type: object
          example: https://example.com/pack.jpg
        priceMode:
          type: string
          enum: &a47
            - FIXED
            - SUM_ITEMS
            - SUM_ITEMS_WITH_DISCOUNT
          example: FIXED
        fixedPrice:
          type: object
          example: 299
        discountAmount:
          type: object
          example: 20
        discountPercentage:
          type: object
          example: 10
        minBudget:
          type: object
          example: 200
        maxBudget:
          type: object
          example: 350
        currency:
          type: string
          example: MAD
          default: MAD
        priority:
          type: number
          example: 5
          default: 0
        status:
          type: string
          enum: &a48
            - DRAFT
            - ACTIVE
            - ARCHIVED
          default: DRAFT
        isActive:
          type: boolean
          example: true
          default: true
        items:
          type: array
          items:
            $ref: "#/components/schemas/PackItemInputDto"
        attributes:
          type: array
          items:
            $ref: "#/components/schemas/PackAttributeInputDto"
      required:
        - name
        - slug
        - priceMode
    UpdatePackDto:
      type: object
      properties:
        name:
          type: string
          example: Natural Glow Pack
        slug:
          type: string
          example: natural-glow-pack
        description:
          type: object
          example: Natural makeup pack
        mainImageUrl:
          type: object
          example: https://example.com/pack.jpg
        priceMode:
          type: string
          enum: *a47
          example: FIXED
        fixedPrice:
          type: object
          example: 299
        discountAmount:
          type: object
          example: 20
        discountPercentage:
          type: object
          example: 10
        minBudget:
          type: object
          example: 200
        maxBudget:
          type: object
          example: 350
        currency:
          type: string
          example: MAD
          default: MAD
        priority:
          type: number
          example: 5
          default: 0
        status:
          type: string
          enum: *a48
          default: DRAFT
        isActive:
          type: boolean
          example: true
          default: true
        items:
          type: array
          items:
            $ref: "#/components/schemas/PackItemInputDto"
        attributes:
          type: array
          items:
            $ref: "#/components/schemas/PackAttributeInputDto"
...